scheduler/requirements.txt
requests
psycopg2-binary
schedule
python-dotenv
websocket-client
